drivers: Reorder subdirectories
The contents of the drivers directory is included in the no-os static
library. The UART API defines no_os_uart_stdio() as weak and each
platform implements it. The static library is created by invoking gcc
and ar, so the linker dosen't resolve weak and strong symbols (thus both
will be included in the library). Whenever linking the library in
another project, the linker will pick the first symbol it finds (the
weak one, since api/ was compiled first) and ignore the other one.

This is a temporary workaround. In the end we probably need to run the
linker on the object files included in the library.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>