drivers: adc: ad463x: Drop < 0 check over unsigned int
It doesn't make sense to have a negative gain over the ADC input signal.
Signal gains are always expressed as a positive values. Thus, the parameter
to indicate how much gain to apply is of unsigned integer type. Because of
that, the gain never contains a negative value and the 'gain < 0' check is
meaningless. Drop < 0 check over unsigned variable also solves a build
warning that complains about the issue describe above.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>